Adds game processing and stats recording
Implements a game processing service to parse demo files
and record game statistics.

Introduces a job for scheduled match synchronization.

Makes team id nullable and converts game stats properties to double for precision.

+++ src/Application/GameStats/Update/UpdateGameStatCommand.cs
@@ -8,11 +8,11 @@ public sealed class UpdateGameStatCommand : ICommand
     public int Kills { get; set; }
     public int Assists { get; set; }
     public int Deaths { get; set; }
-    public int Adr { get; set; }
-    public int HsRatio { get; set; }
+    public double Adr { get; set; }
+    public double HsRatio { get; set; }
     public double ClutchRatio { get; set; }
-    public int Fkpr { get; set; }
-    public int Fdpr { get; set; }
+    public double Fkpr { get; set; }
+    public double Fdpr { get; set; }
     public int RoundsPlayed { get; set; }
     public int TotalDamage { get; set; }
     public int HeadshotKills { get; set; }
